Arithmetic: new arcsecsq-to-pixnum operator
Until now, there was not an operator for converting arcsec2(area) to number of pixels, based on the input of wcs.

With this commit, Gnuastro now has the 'arcsecsq-to-pixnum' operator for doing this conversion. The 'arcsecsq-to-pixnum' operator takes an input area in arcsec^2 and pixelscale from  input WCS file, and returns the number of pixels that region covers.